+++ src/helpers/FetchData.ts
@@ -1,5 +1,5 @@
 export const fetchData = async () => {
-  const res = await fetch("https://6523ab0aea560a22a4e8a349.mockapi.io/menus");
+  const res = await fetch("https://6523ab0aea560a22a4e8a349.mockapi.io/menu");
   if (res.ok) {
     const data = await res.json();
     return data;
